Add files via upload
In the BorrowLoan case, we first check that the account passed in is owned by the program, and then check that the loan account has been initialized. We use the Rent sysvar to ensure that the borrower's account has enough rent-exempt space to store the account data, and then check that the account has enough lamports to cover the borrower_amount. We then transfer the collateral from the loan account to the borrower's account, and transfer the borrowed amount from the loan account to the borrower's account. Finally, we update the loan account with the borrower's public key and set is_initialized to false.